Renaming and corresponding adjustments
